# Website/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentilytics",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login For Access Token",
        "operationId": "login_for_access_token_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_v1_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze": {
      "post": {
        "tags": [
          "analyze"
        ],
        "summary": "Analyze Text",
        "description": "Single-text synchronous analysis.",
        "operationId": "analyze_text_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch": {
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Batch Upload",
        "description": "Bulk file upload (async).",
        "operationId": "batch_upload_v1_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_batch_upload_v1_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Status",
        "description": "Job status & progress snapshot.",
        "operationId": "get_job_status_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/results": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Results",
        "description": "Paginated row-level results.",
        "operationId": "get_job_results_v1_jobs__job_id__results_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/export": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Export Job Results",
        "description": "Streamed CSV download.",
        "operationId": "export_job_results_v1_jobs__job_id__export_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/stream": {
      "get": {
        "tags": [
          "stream"
        ],
        "summary": "Job Stream",
        "description": "Live progress via SSE.",
        "operationId": "job_stream_v1_jobs__job_id__stream_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "operationId": "health_check_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisResultResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "row_index": {
            "type": "integer",
            "title": "Row Index"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "compound_score": {
            "type": "number",
            "title": "Compound Score"
          },
          "positive_score": {
            "type": "number",
            "title": "Positive Score"
          },
          "neutral_score": {
            "type": "number",
            "title": "Neutral Score"
          },
          "negative_score": {
            "type": "number",
            "title": "Negative Score"
          },
          "overall_sentiment": {
            "type": "string",
            "title": "Overall Sentiment"
          },
          "aspects": {
            "items": {
              "$ref": "#/components/schemas/AspectDetail"
            },
            "type": "array",
            "title": "Aspects"
          },
          "key_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Phrases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "row_index",
          "raw_text",
          "compound_score",
          "positive_score",
          "neutral_score",
          "negative_score",
          "overall_sentiment",
          "aspects",
          "key_phrases"
        ],
        "title": "AnalysisResultResponse"
      },
      "AnalyzeMetadata": {
        "properties": {
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms"
          },
          "llm_used": {
            "type": "boolean",
            "title": "Llm Used"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "processing_time_ms",
          "llm_used"
        ],
        "title": "AnalyzeMetadata"
      },
      "AnalyzeOptions": {
        "properties": {
          "include_aspects": {
            "type": "boolean",
            "title": "Include Aspects",
            "default": true
          },
          "include_key_phrases": {
            "type": "boolean",
            "title": "Include Key Phrases",
            "default": true
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "title": "AnalyzeOptions"
      },
      "AnalyzeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyzeOptions"
              },
              {
                "type": "null"
              }
            ],
            "default": {
              "include_aspects": true,
              "include_key_phrases": true,
              "language": "en"
            }
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "overall_sentiment": {
            "$ref": "#/components/schemas/SentimentScore"
          },
          "aspects": {
            "items": {
              "$ref": "#/components/schemas/AspectDetail"
            },
            "type": "array",
            "title": "Aspects"
          },
          "key_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Phrases"
          },
          "metadata": {
            "$ref": "#/components/schemas/AnalyzeMetadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "overall_sentiment",
          "aspects",
          "key_phrases",
          "metadata"
        ],
        "title": "AnalyzeResponse"
      },
      "AspectDetail": {
        "properties": {
          "aspect": {
            "type": "string",
            "title": "Aspect"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "sentiment": {
            "type": "string",
            "title": "Sentiment"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "aspect",
          "category",
          "sentiment",
          "confidence",
          "evidence"
        ],
        "title": "AspectDetail"
      },
      "Body_batch_upload_v1_batch_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "column_mapping": {
            "type": "string",
            "title": "Column Mapping"
          },
          "options": {
            "type": "string",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "file",
          "column_mapping",
          "options"
        ],
        "title": "Body_batch_upload_v1_batch_post"
      },
      "Body_login_for_access_token_v1_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_for_access_token_v1_auth_token_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "processed_rows": {
            "type": "integer",
            "title": "Processed Rows"
          },
          "failed_rows": {
            "type": "integer",
            "title": "Failed Rows"
          },
          "estimated_completion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Completion"
          },
          "summary_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary Distribution"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "total_rows",
          "processed_rows",
          "failed_rows",
          "summary_distribution"
        ],
        "title": "JobStatusResponse"
      },
      "PaginatedResults": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisResultResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedResults"
      },
      "SentimentScore": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "positive_score": {
            "type": "number",
            "title": "Positive Score"
          },
          "neutral_score": {
            "type": "number",
            "title": "Neutral Score"
          },
          "negative_score": {
            "type": "number",
            "title": "Negative Score"
          },
          "compound_score": {
            "type": "number",
            "title": "Compound Score"
          }
        },
        "type": "object",
        "required": [
          "label",
          "positive_score",
          "neutral_score",
          "negative_score",
          "compound_score"
        ],
        "title": "SentimentScore"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "token_type"
        ],
        "title": "Token"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}